refactor(ui): change FileUpload from atom to molecule
re #41
